rename assets.ts -> useAssets.ts
