+++ src/sections/detailsDrawer/AssetDrawer.tsx
@@ -11,7 +11,7 @@ import { Tooltip } from '@/components/Tooltip';
 import { AssetStatusChip } from '@/components/ui/AssetStatusChip';
 import { DetailsListContainer } from '@/components/ui/DetailsListContainer';
 import { useMy } from '@/hooks';
-import { useUpdateAsset } from '@/hooks/assets';
+import { useUpdateAsset } from '@/hooks/useAssets';
 import { useGenericSearch } from '@/hooks/useGenericSearch';
 import { useSearchParams } from '@/hooks/useSearchParams';
 import { formatDate } from '@/utils/date.util';
